pyproject.toml
[project]
name = "my-grader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.2.1",
    "alembic>=1.17.0",
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "environ>=1.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.120.1",
    "openai>=2.6.1",
    "pypdf>=6.1.3",
    "pypdf2>=3.0.1",
    "sentence-transformers>=5.1.2",
    "uuid>=1.30",
    "uvicorn>=0.38.0",
]
